ENH: Add a deformable motion trajectory
This commit adds a new TrajectoryDeformation class to support non-rigid
motion of samples. This class loads an HDF5 file which contains a 4D
deformation field grid of (dx, dy, dz) triples, along with a vector of
scales for time, x, y and z in ms and mm respectively. The initial
position of each spin is used to index into this grid and quad-linear
interpolation is used to find the deformation of the given spin, which
is then applied to the spin's position.

There is also a small modification of the NDData class to allow the
indexing of 5 dimensional grids, as this is the underlying storage
used for the deformation field in the class.

This code is not suitable as is for inclusion in the master version of
JEMRIS, as it currently replaces the standard rigid motion with this
non-rigid motion. Some discussion will be necessary to decide on the
best way to integrate this code to work alongside the existing code.